implemented methods for aggregation
