+++ package.json
@@ -12,8 +12,8 @@
     "devDependencies": {
         "@energyleaf/tsconfig": "workspace:*",
         "@total-typescript/ts-reset": "^0.5.1",
-        "eslint": "^8.52.0",
-        "prettier": "^3.0.3",
+        "eslint": "^8.54.0",
+        "prettier": "^3.1.0",
         "turbo": "^1.10.16"
     },
     "packageManager": "pnpm@8.6.10",
@@ -21,7 +21,7 @@
     "dependencies": {
         "@ianvs/prettier-plugin-sort-imports": "^4.1.1",
         "dotenv-cli": "^7.3.0",
-        "prettier-plugin-tailwindcss": "^0.5.6",
-        "typescript": "^5.2.2"
+        "prettier-plugin-tailwindcss": "^0.5.7",
+        "typescript": "^5.3.2"
     }
 }
